requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
python-dotenv
anthropic
z3-solver
pinecone
openai
requests
tqdm
tokenizers